+++ server/routes/userRoute.js
@@ -1,16 +1,15 @@
 const express = require("express");
 const router = express.Router();
 const controlRoute = require("../controllers/user");
-const authController = require('../controllers/authController')
 
 router.post('/signup',controlRoute.signup_post);
 
 router.get('/signup',controlRoute.signup_usr);
 
 // Getting the login page
-router.get('/login', authController.login_get)
+router.get('/login', controlRoute.login_get)
 
 //Sending the login information for verification
-router.post('/login', authController.login_post)
+router.post('/login', controlRoute.login_post)
 
 module.exports = router;
